hashes: Rename util to macros
The `utils` module holds public macros, call it `macros` instead.
