+++ hashes/src/lib.rs
@@ -96,7 +96,7 @@ pub mod _export {
 
 mod internal_macros;
 #[macro_use]
-mod util;
+mod macros;
 
 #[macro_use]
 pub mod serde_macros;
