+++ hashes/src/macros.rs
